Git switching from internal to external modules.
