+++ src/linq/iterables/takeWhile.ts
@@ -1,0 +1,18 @@
+export class TakeWhile<TSource>  {
+    private _elements: Iterable<TSource>;
+    private _predicate: (element: TSource) => boolean;
+
+    constructor(elements: Iterable<TSource>, predicate: (element: TSource) => boolean) {
+        this._elements = elements;
+        this._predicate = predicate;
+    }
+
+    *[Symbol.iterator](): Iterator<TSource> {
+        for (let element of this._elements) {
+            if (!this._predicate(element)) {
+                break;
+            }
+            yield element;
+        }
+    }
+}
