+++ src/linq/list.ts
@@ -1,14 +1,11 @@
-namespace Linq {
-    export class List<TSource> extends Linqable<TSource> {
-        private _elements: Iterable<any>;
+export class List<TSource>  {
+    private _elements: Iterable<any>;
 
-        constructor(elements: Iterable<any>) {
-            super();
-            this._elements = elements;
-        }
+    constructor(elements: Iterable<any>) {
+        this._elements = elements;
+    }
 
-        *[Symbol.iterator](): Iterator<TSource> {
-            yield* this._elements;
-        }
+    *[Symbol.iterator](): Iterator<TSource> {
+        yield* this._elements;
     }
 }
